Remove bad cast

+++ test/hotspot/gtest/oops/test_objArrayOop.cpp
@@ -63,7 +63,7 @@ TEST_VM(objArrayOop, osize) {
   };
   for (int i = 0; x[i].result != -1; i++) {
     if (x[i].objal == (int)ObjectAlignmentInBytes && x[i].ccp == UseCompressedClassPointers && x[i].coops == UseCompressedOops && x[i].coh == UseCompactObjectHeaders) {
-      EXPECT_EQ(objArrayOopDesc::object_size(1), (size_t)x[i].result);
+      EXPECT_EQ(objArrayOopDesc::object_size(1), x[i].result);
     }
   }
 }
